Update ABI mode documentation
We have just added extra functionality to this abi mode. We are now able to ignore certain abi and address combinations, in the event  that they do now return what we want. This will greatly improve performance of self discovery because in most cases abi and addresses are not compatible. This prevents the system having to instantiate web3 contract instances for abi and addresses which are not (and as such will never be) associated.